backend/requirements.txt
asgiref==3.7.2
blinker==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.1.0
psycopg==3.1.13
psycopg-binary==3.1.13
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
